requirements2.txt
langchain
cohere
openai
python-dotenv
streamlit
huggingface_hub